Preventation of deleting added, alert for succesfully updating added, inoices tab deleted

+++ public/js/pages/customers/customersController.js
@@ -5,13 +5,20 @@ app.controller('customersCntrl', ["$scope", "$http","REST_CONST", "$rootScope", 
         }
     );
     $scope.removeCustomerFromTable= function (customerId) {
-        $http.delete(REST_CONST.url + '/customers/' + customerId).then(function (resp) {
-            console.log('Deleted customer with id '+ resp.data.id);
-            $http.get(REST_CONST.url + '/customers').then(function (resp) {
-                    $rootScope.customerData = resp.data;
-                }
-            )
-        })
+        var confirm = window.confirm("Are you sure want to delete this customer?");
+        console.log(confirm);
+        if(confirm) {
+            $http.delete(REST_CONST.url + '/customers/' + customerId).then(function (resp) {
+                console.log('Deleted customer with id ' + resp.data.id);
+                $http.get(REST_CONST.url + '/customers').then(function (resp) {
+                        $rootScope.customerData = resp.data;
+                    }
+                )
+            })
+        } else {
+            console.log("Deleting prevented from console");
+            return false
+        }
     }
 
 }]);
